calendar_이지연_240220

+++ src/components/fullcalendar/CommonCalendarModal.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Modal, Button, Form } from 'react-bootstrap';
 import moment from 'moment-timezone';
 
-const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => {
+const CommonCalendarModal = ({ action, event, onSave, onUpdate, onDelete, onClose, categories }) => {
     // 날짜 형식을 변환하는 함수
     const formatDateForInput = (dateString) => {
         const momentDate = moment.tz(dateString, 'Asia/Seoul');
@@ -22,12 +22,12 @@ const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => 
         if (event) {
             // 이벤트 클릭 시 -> update, delete
             setFormData({
-                title: event.title,
-                start: formatDateForInput(event.start),
-                end: formatDateForInput(event.end),
-                no: event.extendedProps.no,
-                category: event.extendedProps.category,
-                content: event.extendedProps.content,
+                title: event.title || '',
+                start: formatDateForInput(event.start) || '',
+                end: formatDateForInput(event.end) || '',
+                no: event.extendedProps?.no || '', 
+                category: event.extendedProps?.category || '', // 수정된 부분
+                content: event.extendedProps?.content || '',
             });
         } else {
             // 날짜 클릭 시 -> create
@@ -52,33 +52,43 @@ const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => 
 
     //저장하기
     const handleSave = () => {
-        onSave(formData);
+        const confirmAdd = window.confirm("생성하시겠습니까?");
+        if(confirmAdd){
+            onSave(formData);
+            onClose();
+        }
         onClose();
     };
 
     //수정하기
     const handleUpdate = () => {
-        onUpdate(formData);
+        const confirmUpdate = window.confirm("수정하시겠습니까?");
+        if(confirmUpdate){
+            onUpdate(formData);
+            onClose();
+        }
         onClose();
     };
     
     //삭제하기
     const handleDelete = () => {
-        // 삭제 로직 수행
-        onDelete(formData);
-        console.log(formData);
+        const confirmDelete = window.confirm("삭제하시겠습니까?");
+        if(confirmDelete){
+            onDelete(formData);
+            onClose();
+        }
         onClose();
     };
 
     return (
         <Modal show={true} onHide={onClose}>
-            <Modal.Header closeButton>
-                <Modal.Title>{action === '생성' ? 'New 일정 추가' : '일정 수정'}</Modal.Title>
+            <Modal.Header closeButton style={{backgroundColor:'#E6E6FA'}}>
+                <Modal.Title style={{fontWeight:'bolder'}}>{action === '생성' ? '일정 추가' : '일정 수정'}</Modal.Title>
             </Modal.Header>
             <Modal.Body>
                 <Form>
                     <Form.Group controlId="formTitle">
-                        <Form.Label>일정 이름</Form.Label>
+                        <Form.Label style={{fontSize:'20px', margin:'15px', fontWeight:'bolder'}}>일정 이름</Form.Label>
                         <Form.Control
                             type="text"
                             placeholder="일정 이름을 입력하세요"
@@ -88,17 +98,34 @@ const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => 
                         />
                     </Form.Group>
                     <Form.Group controlId="formContent">
-                        <Form.Label>내용</Form.Label>
+                        <Form.Label style={{fontSize:'20px', margin:'15px', fontWeight:'bolder'}}>내용</Form.Label>
                         <Form.Control
                             as="textarea"
-                            rows={3}
+                            rows={1}
+                            style={{ height: '80px' }}  // 원하는 높이로 조절
                             name="content"
                             value={formData.content}
                             onChange={handleChange}
                         />
                     </Form.Group>
+                    <Form.Group controlId="formCategory">
+                        <Form.Label style={{fontSize:'20px', margin:'15px', fontWeight:'bolder'}}>카테고리</Form.Label>
+                        <Form.Control
+                            as="select"
+                            name="category"
+                            value={formData.category}
+                            onChange={handleChange}
+                        >
+                            <option value="" >카테고리를 선택하세요</option>
+                            {categories.map((category) => (
+                                <option key={category} value={category}>
+                                    {category}
+                                </option>
+                            ))}
+                        </Form.Control>
+                    </Form.Group>
                     <Form.Group controlId="formStart">
-                        <Form.Label>시작 일시</Form.Label>
+                        <Form.Label style={{fontSize:'20px', margin:'15px', fontWeight:'bolder'}}>시작 일시</Form.Label>
                         <Form.Control
                             type="datetime-local"
                             name="start"
@@ -106,9 +133,8 @@ const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => 
                             onChange={handleChange}
                             />
                     </Form.Group>
-                    <br/>
                     <Form.Group controlId="formEnd">
-                        <Form.Label>종료 일시</Form.Label>
+                        <Form.Label style={{fontSize:'20px', margin:'15px', fontWeight:'bolder'}}>종료 일시</Form.Label>
                         <Form.Control
                             type="datetime-local"
                             name="end"
@@ -119,15 +145,15 @@ const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => 
                     <br/>
                 </Form>
             </Modal.Body>
-            <Modal.Footer>
-                <Button variant="secondary" onClick={onClose}>
+            <Modal.Footer style={{backgroundColor:'#E6E6FA'}}>
+                <Button style={{fontWeight:'bolder'}} variant="outline-secondary" onClick={onClose}>
                     취소
                 </Button>
-                <Button variant="primary" onClick={action === '생성' ? handleSave : handleUpdate}>
+                <Button style={{fontWeight:'bolder'}} variant="outline-primary"  onClick={action === '생성' ? handleSave : handleUpdate}>
                     {action === '생성' ? '저장' : '수정'}
                 </Button>
                 {action === '수정' && (
-                    <Button variant="danger" onClick={handleDelete}>
+                    <Button style={{fontWeight:'bolder'}} variant="outline-danger" onClick={handleDelete}>
                         삭제
                     </Button>
                 )}
@@ -136,4 +162,4 @@ const CustomModal = ({ action, event, onSave, onUpdate, onDelete, onClose }) => 
     );
 };
 
-export default CustomModal;
+export default CommonCalendarModal;
